requirements.txt
# requirements.txt
requests>=2.28.0  #significa: “Instalá la versión 2.28.0 o una más nueva compatible”
pandas>=1.5.0
pyarrow>=10.0.0
python-dotenv>=0.21.0